+++ CHANGELOG.md
@@ -7,6 +7,7 @@ We follow [Semantic Versions](https://semver.org/).
 - Make `Dependency.__call__` async because FastAPI always runs Picodi deps in threadpool even if they are async
 - Rename `Dependency` to `Depends` [internal change]
 - Run mypy in tests directory
+- Make sure that `inject` doesn't change type of wrapped function (e.g. coroutinefunction, generatorfunction, etc.)
 
 ## Version 0.20.0