+++ CHANGELOG.md
@@ -4,6 +4,10 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Version next
 
+-
+
+## Version 0.21.0
+
 - Make `Dependency.__call__` async because FastAPI always runs Picodi deps in threadpool even if they are async
 - Rename `Dependency` to `Depends` [internal change]
 - Run mypy in tests directory